chor: create utils file for tree
